Updates to settings to allow for arm64 builds.  This was possible prior but now configured so by default.  Added casts to integer to string conversions.  Changed two-part initializations to use first-part expand properties.

+++ Version3/Sources/MASTAdView/MASTAdView/MASTMRAIDExpandProperties.m
@@ -60,7 +60,8 @@ static NSString* MASTMRAIDExpandPropertiesUseCustomClose = @"useCustomClose";
     if (self.useCustomClose)
         ucc = @"true";
     
-    NSString* desc = [NSString stringWithFormat:@"{width:%d,height:%d,useCustomClose:%@}", self.width, self.height, ucc];
+    NSString* desc = [NSString stringWithFormat:@"{width:%d,height:%d,useCustomClose:%@}",
+                      (int)self.width, (int)self.height, ucc];
     
     return desc;
 }
